Add remaining settings
