+++ src/Admin/Settings/General.php
@@ -1,4 +1,10 @@
 <?php
+/**
+ * The general admin settings.
+ *
+ * @package WCVendors/Admin/Settings
+ * @phpcs:disable WordPress.WP.I18n
+ */
 
 namespace WCVendors\Admin\Settings;
 
@@ -7,20 +13,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 /**
- * The general admin settings
- *
- * @author      Jamie Madden, WC Vendors
- * @category    Settings
- * @package     WCVendors/Admin/Settings
- * @version     2.0.0
- */
-
-if ( ! defined( 'ABSPATH' ) ) {
-	exit; // Exit if accessed directly
-}
-
-/**
- * WC_Admin_Settings_General.
+ * The general admin settings.
  */
 class General extends Base {
 
@@ -50,58 +43,53 @@ class General extends Base {
 	/**
 	 * Get settings array.
 	 *
+	 * @param string $current_section Current section ID.
+	 *
 	 * @return array
 	 */
 	public function get_settings( $current_section = '' ) {
 
-		$settings = apply_filters(
-			'wcvendors_settings',
+		$settings = array(
+			// General Options.
 			array(
-
-				// General Options
-				array(
-					'type' => 'title',
-					'desc' => __( 'These are the general settings for your marketplace', 'wc-vendors' ),
-					'id'   => 'general_options',
-				),
-				array(
-					'title'   => sprintf( __( '%s Registration', 'wc-vendors' ), wcv_get_vendor_name() ),
-					'desc'    => sprintf( __( 'Allow users to apply to become a %s', 'wc-vendors' ), lcfirst( wcv_get_vendor_name() ) ),
-					'id'      => 'wcvendors_vendor_allow_registration',
-					'default' => 'no',
-					'type'    => 'checkbox',
-				),
-				array(
-					'title'   => sprintf( __( '%s Approval', 'wc-vendors' ), wcv_get_vendor_name() ),
-					'desc'    => sprintf( __( 'Manually approve all %s applications', 'wc-vendors' ), lcfirst( wcv_get_vendor_name() ) ),
-					'id'      => 'wcvendors_vendor_approve_registration',
-					'default' => 'no',
-					'type'    => 'checkbox',
-				),
-
-				array(
-					'title'   => sprintf( __( '%s Taxes', 'wc-vendors' ), wcv_get_vendor_name() ),
-					'desc'    => sprintf( __( 'Give any taxes to the %s', 'wc-vendors' ), wcv_get_vendor_name() ),
-					'id'      => 'wcvendors_vendor_give_taxes',
-					'default' => 'no',
-					'type'    => 'checkbox',
-				),
-				array(
-					'title'   => sprintf( __( '%s Shipping', 'wc-vendors' ), wcv_get_vendor_name() ),
-					'desc'    => sprintf( __( 'Give any shipping to the %s', 'wc-vendors' ), wcv_get_vendor_name() ),
-					'id'      => 'wcvendors_vendor_give_shipping',
-					'default' => 'no',
-					'type'    => 'checkbox',
-				),
-				array(
-					'type' => 'sectionend',
-					'id'   => 'general_options',
-				),
-
-			)
+				'type' => 'title',
+				'desc' => __( 'These are the general settings for your marketplace', 'wc-vendors' ),
+				'id'   => 'general_options',
+			),
+			array(
+				'title'   => sprintf( __( '%s Registration', 'wc-vendors' ), wcv_get_vendor_name() ),
+				'desc'    => sprintf( __( 'Allow users to apply to become a %s', 'wc-vendors' ), lcfirst( wcv_get_vendor_name() ) ),
+				'id'      => 'wcvendors_vendor_allow_registration',
+				'default' => 'no',
+				'type'    => 'checkbox',
+			),
+			array(
+				'title'   => sprintf( __( '%s Approval', 'wc-vendors' ), wcv_get_vendor_name() ),
+				'desc'    => sprintf( __( 'Manually approve all %s applications', 'wc-vendors' ), lcfirst( wcv_get_vendor_name() ) ),
+				'id'      => 'wcvendors_vendor_approve_registration',
+				'default' => 'no',
+				'type'    => 'checkbox',
+			),
+			array(
+				'title'   => sprintf( __( '%s Taxes', 'wc-vendors' ), wcv_get_vendor_name() ),
+				'desc'    => sprintf( __( 'Give any taxes to the %s', 'wc-vendors' ), wcv_get_vendor_name() ),
+				'id'      => 'wcvendors_vendor_give_taxes',
+				'default' => 'no',
+				'type'    => 'checkbox',
+			),
+			array(
+				'title'   => sprintf( __( '%s Shipping', 'wc-vendors' ), wcv_get_vendor_name() ),
+				'desc'    => sprintf( __( 'Give any shipping to the %s', 'wc-vendors' ), wcv_get_vendor_name() ),
+				'id'      => 'wcvendors_vendor_give_shipping',
+				'default' => 'no',
+				'type'    => 'checkbox',
+			),
+			array(
+				'type' => 'sectionend',
+				'id'   => 'general_options',
+			),
 		);
 
 		return apply_filters( 'wcvendors_get_settings_' . $this->id, $settings );
 	}
-
 }
